Add popup skipping when not found, add debug parameter to driver
Popup skipping is risky because the popups may not be found due to
an error and not because they do not exist. Must add click interception
handling in next clicks so that it retries closing the popups.

Added new savePageSource method in driver for writing the current page's
source to a file.