Driver closes popups only when needed, retries closing if there is problem
During the driver setup the driver will visit the main aliexpress page
and close the popups only when there is a country and/or a currency to
setup.

If a popup is skipped when it shouldn't and later intercepts a click on
the settings menu, the driver tries to close it once more. If the click
is then intercepted again an exception is left to be raised.